image dynamic

+++ src/Page/AddInventoryProduct.jsx
@@ -34,18 +34,11 @@ function AddInventoryProduct() {
     const form = e.target;
     const category = parseInt(form.category.value);
     const itemName = form.itemName.value;
-    const otherCost = parseInt(form.otherCost.value);
-    const inventoryCost = parseInt(form.inventoryCost.value);
-    const transportationCost = parseInt(form.transportationCost.value);
+    const otherCost = parseFloat(form.otherCost.value);
+    const inventoryCost = parseFloat(form.inventoryCost.value);
+    const transportationCost = parseFloat(form.transportationCost.value);
     const unit = parseInt(form.unit.value);
-    // const data = {
-    //   category,
-    //   itemName,
-    //   otherCost,
-    //   inventoryCost,
-    //   transportationCost,
-    //   unit,
-    // };
+
     const formData = new FormData();
     formData.append('category', category);
     formData.append('itemName', itemName);
